Form, projects and side nav CSS styles
Applied styles to the form submit button in the form section of the homepage. Also applied a hover effect and media queries to the button to be more in line with the portfolio concept. Applied more styles to the section. The form containers now have a box shadow. Reworked the styles applied to the projects section. Box shadow is applied to the project containers and removed unnecessary HTML and CSS. Applied more styles to the side nav par to be more in line with the portfolio concept.