Coding examples rework
Reworked the coding examples page. Implemented a code block API into the project that creates code blocks for each code snippet on the coding examples page. Added coding pills to the language's used section in each code snippet to be more in line with the rest of the page. applied multiple media queries to the coding snippets to make the page more responsive and flow better across different view ports. Tested the page for error's, none found. Uploading changes to GitHub and cPanel.