Refactoring.
increased the padding value of section and content containers to 15px for improved spacing and readability. Removed top padding from project items on the index page to sit correctly in their respective cards. Applied the same section styling to the contact form to be inline with the rest of the website. Moved the get in touch section to sit above the contact form at all viewports for easier user navigation and readability.