about me rework.
Moved the about me section to the index page. Reworked the styling to be inline with the rest of the page. Combined the about me and skill containers into one to improve the sections flow and readability. Changed the about me link on all navbars to redirect to the about me section on the index page.